Task 5

+++ 0x0C-more_malloc_free/101-mul.c
@@ -1,5 +1,5 @@
-#include <stdio.h>
 #include "main.h"
+#include <stdio.h>
 
 /**
  * main - program that multiplies two positive numbers.
@@ -11,25 +11,26 @@
 int main(int argc, char *argv[])
 {
 unsigned long mul;
-	int i, j;
+int i, j;
 
 if (argc != 3)
 {
-	printf("Error\n");
-	exit(98);
+printf("Error\n");
+exit(98);
 }
+
 for (i = 1; i < argc; i++)
 {
 for (j = 0; argv[i][j] != '\0'; j++)
 {
 if (argv[i][j] > 57 || argv[i][j] < 48)
 {
-	printf("Error\n");
-	exit(98);
+printf("Error\n");
+exit(98);
 }
 }
 }
-mul = atol(argv[1]) * atol(argv[2]);
+mul = (atoi(argv[1]) * atoi(argv[2]));
 printf("%lu\n", mul);
 
 return (0);
